requirements.txt
streamlit==1.26.0
pandas==2.0.3
python-dotenv==1.0.0
langchain-groq==0.1.2
langchain==0.1.16
langchain-core==0.1.42
plotly==5.18.0
matplotlib==3.8.4
seaborn==0.13.2
openpyxl==3.1.2

-e .